+++ config/urls.py
@@ -12,6 +12,7 @@ urlpatterns = [
     path("profiles/", include("repo.profiles.urls")),
     path("records/", include("repo.records.urls")),
     path("recommendation/", include("repo.recommendation.urls")),
+    path("interactions/", include("repo.interactions.urls")),
     # def-spectacular
     path("api/schema/", SpectacularAPIView.as_view(), name="schema"),
     path("api/docs/", SpectacularSwaggerView.as_view(url_name="schema"), name="swagger-ui"),
