+++ repo/interactions/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from repo.interactions.models import Relationship
+
+admin.site.register(Relationship)
