+++ repo/profiles/urls.py
@@ -14,10 +14,6 @@ urlpatterns = [
     path("signup/", views.SignupView.as_view(), name="signup"),
     path("", views.MyProfileAPIView.as_view(), name="my_profile"),
     path("<int:id>/", views.OtherProfileAPIView.as_view(), name="other_profile"),
-    path("<int:id>/follow/", views.FollowListCreateDeleteAPIView.as_view(), name="follow"),
-    path("follow/", views.FollowListAPIView.as_view(), name="my_follow_list"),
-    path("<int:id>/block/", views.BlockListCreateDeleteAPIView.as_view(), name="my_block_list"),
-    path("block/", views.BlockListAPIView.as_view(), name="my_block_list"),
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
     path("<int:id>/posts/", views.UserPostListAPIView.as_view(), name="user_posts"),
