+++ repo/records/services.py
@@ -15,7 +15,7 @@ from django.shortcuts import get_object_or_404
 from django.utils import timezone
 
 from repo.common.view_counter import is_viewed
-from repo.profiles.models import Relationship
+from repo.interactions.models import Relationship
 from repo.records.models import Comment, Photo, Post, TastedRecord
 from repo.records.posts.serializers import PostListSerializer
 from repo.records.tasted_record.serializers import TastedRecordListSerializer
